refactor(python): move writers + graph adapters into infomap.io package
git mv _writers/_networkx/_igraph/_scipy/_edge_index/export into infomap/io/
(writers, networkx, igraph, scipy, edge_index, export). _facade imports updated
to the io.* paths; the two adapters that needed Infomap now use '..\_facade'.
Optional third-party imports stay lazy (inside functions), so importing infomap
still pulls in no networkx/igraph/scipy/pandas.

A back-compat shim infomap/export.py re-exports the public export helpers
(annotate_networkx_graph, annotate_igraph_graph, write_graphml, write_gexf) from
infomap.io.export so 'from infomap.export import ...' keeps working
(test_export.py). test_public_surface OWNED_MODULES: infomap._writers ->
infomap.io.writers; the pinned signatures are unchanged so the baseline JSON is
untouched. test_igraph/test_scipy/test_edge_index updated to the io.* paths.
