+++ interfaces/python/src/infomap/io/edge_index.py
