+++ interfaces/python/src/infomap/io/scipy.py
