+++ interfaces/python/src/infomap/io/writers.py
